Minor refactor in DNS tests
