+++ dns_server/src/crawler/tests/mod.rs
@@ -13,32 +13,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    net::{IpAddr, SocketAddr},
-    sync::{Arc, Mutex},
-    time::Duration,
-};
+mod mock;
 
-use async_trait::async_trait;
-use common::primitives::semver::SemVer;
-use p2p::{
-    config::P2pConfig,
-    error::{DialError, P2pError},
-    message::{
-        AnnounceAddrRequest, Announcement, PeerManagerRequest, PeerManagerResponse, SyncRequest,
-        SyncResponse,
-    },
-    net::{
-        default_backend::{
-            transport::TransportAddress,
-            types::{PeerId, RequestId},
-        },
-        types::{ConnectivityEvent, PeerInfo, SyncingEvent},
-        ConnectivityService, NetworkingService, SyncingMessagingService,
-    },
-    testing_utils::P2pTestTimeGetter,
-};
+use std::{net::SocketAddr, time::Duration};
+
+use p2p::testing_utils::P2pTestTimeGetter;
 use tokio::sync::mpsc;
 
 use crate::{
@@ -49,161 +28,11 @@ use crate::{
     dns_server::ServerCommands,
 };
 
+use self::mock::{
+    MockConnectivityHandle, MockNetworkingService, MockStateRef, MockSyncingMessagingHandle,
+};
+
 use super::storage_impl::DnsServerStorageImpl;
-
-#[derive(Clone)]
-struct MockStateRef {
-    crawler_config: CrawlerConfig,
-    online: Arc<Mutex<BTreeSet<SocketAddr>>>,
-    connected: Arc<Mutex<BTreeMap<SocketAddr, PeerId>>>,
-    connection_attempts: Arc<Mutex<Vec<SocketAddr>>>,
-    conn_tx: mpsc::UnboundedSender<ConnectivityEvent<MockNetworkingService>>,
-}
-
-impl MockStateRef {
-    fn node_online(&self, ip: SocketAddr) {
-        let added = self.online.lock().unwrap().insert(ip);
-        assert!(added);
-    }
-
-    fn node_offline(&self, ip: SocketAddr) {
-        let removed = self.online.lock().unwrap().remove(&ip);
-        assert!(removed);
-        if let Some(peer_id) = self.connected.lock().unwrap().remove(&ip) {
-            self.conn_tx.send(ConnectivityEvent::ConnectionClosed { peer_id }).unwrap();
-        }
-    }
-
-    fn announce_address(&self, from: SocketAddr, announced_ip: SocketAddr) {
-        let peer_id = *self.connected.lock().unwrap().get(&from).unwrap();
-        self.conn_tx
-            .send(ConnectivityEvent::Request {
-                peer_id,
-                request_id: RequestId::new(),
-                request: PeerManagerRequest::AnnounceAddrRequest(AnnounceAddrRequest {
-                    address: announced_ip.as_peer_address(),
-                }),
-            })
-            .unwrap();
-    }
-}
-
-#[derive(Debug)]
-struct MockNetworkingService {}
-
-struct MockConnectivityHandle {
-    state: MockStateRef,
-    conn_rx: mpsc::UnboundedReceiver<ConnectivityEvent<MockNetworkingService>>,
-}
-
-struct MockSyncingMessagingHandle {}
-
-#[async_trait]
-impl NetworkingService for MockNetworkingService {
-    type Transport = ();
-    type Address = SocketAddr;
-    type BannableAddress = IpAddr;
-    type PeerId = PeerId;
-    type PeerRequestId = RequestId;
-    type ConnectivityHandle = MockConnectivityHandle;
-    type SyncingMessagingHandle = MockSyncingMessagingHandle;
-
-    async fn start(
-        _transport: Self::Transport,
-        _bind_addresses: Vec<Self::Address>,
-        _chain_config: Arc<common::chain::ChainConfig>,
-        _p2p_config: Arc<P2pConfig>,
-    ) -> p2p::Result<(Self::ConnectivityHandle, Self::SyncingMessagingHandle)> {
-        unreachable!()
-    }
-}
-
-#[async_trait]
-impl ConnectivityService<MockNetworkingService> for MockConnectivityHandle {
-    fn connect(&mut self, address: SocketAddr) -> p2p::Result<()> {
-        self.state.connection_attempts.lock().unwrap().push(address);
-        if self.state.online.lock().unwrap().contains(&address) {
-            let peer_id = PeerId::new();
-            let peer_info = PeerInfo {
-                peer_id,
-                network: self.state.crawler_config.network,
-                version: SemVer::new(1, 2, 3),
-                agent: None,
-                subscriptions: BTreeSet::new(),
-            };
-            let old = self.state.connected.lock().unwrap().insert(address, peer_id);
-            assert!(old.is_none());
-            self.state
-                .conn_tx
-                .send(ConnectivityEvent::OutboundAccepted {
-                    address,
-                    peer_info,
-                    receiver_address: None,
-                })
-                .unwrap();
-        } else {
-            self.state
-                .conn_tx
-                .send(ConnectivityEvent::ConnectionError {
-                    address,
-                    error: P2pError::DialError(DialError::ConnectionRefusedOrTimedOut),
-                })
-                .unwrap();
-        }
-        Ok(())
-    }
-
-    fn disconnect(&mut self, _peer_id: PeerId) -> p2p::Result<()> {
-        unreachable!()
-    }
-
-    fn send_request(
-        &mut self,
-        _peer_id: PeerId,
-        _request: PeerManagerRequest,
-    ) -> p2p::Result<RequestId> {
-        unreachable!()
-    }
-
-    fn send_response(
-        &mut self,
-        _request_id: RequestId,
-        _response: PeerManagerResponse,
-    ) -> p2p::Result<()> {
-        unreachable!()
-    }
-
-    fn local_addresses(&self) -> &[SocketAddr] {
-        unreachable!()
-    }
-
-    async fn poll_next(&mut self) -> p2p::Result<ConnectivityEvent<MockNetworkingService>> {
-        Ok(self.conn_rx.recv().await.unwrap())
-    }
-}
-
-#[async_trait]
-impl SyncingMessagingService<MockNetworkingService> for MockSyncingMessagingHandle {
-    fn send_request(&mut self, _peer_id: PeerId, _request: SyncRequest) -> p2p::Result<RequestId> {
-        unreachable!()
-    }
-
-    fn send_response(
-        &mut self,
-        _request_id: RequestId,
-        _response: SyncResponse,
-    ) -> p2p::Result<()> {
-        unreachable!()
-    }
-
-    fn make_announcement(&mut self, _announcement: Announcement) -> p2p::Result<()> {
-        unreachable!()
-    }
-
-    async fn poll_next(&mut self) -> p2p::Result<SyncingEvent<MockNetworkingService>> {
-        std::future::pending().await
-    }
-}
 
 fn test_crawler(
     add_node: Vec<SocketAddr>,
